Se realizan el cambio de la IU pedido por el profesor, y se onda el tema
de responsividad, con este pull request completamos los aspectos del
segundo parcial.

+++ src/components/ChangeImageProfile.tsx
@@ -2,10 +2,13 @@ import React, { useState, useRef } from "react";
 import './ChangeImageProfile.css';
 import ConfirmationModal from "./ConfirmationModal";
 import { subirImagenPerfil } from "../redux/Trunks/accionesVariasThunk";
-import { useDispatch } from "react-redux";
-import { AppDispatch } from "../redux/store";
+import { useDispatch, useSelector } from "react-redux";
+import { AppDispatch, RootState } from "../redux/store";
 import LoadingModal from "./LoadignModal";
-import { deleteImgProfile, editImgProfile } from "../redux/Trunks/userTrunk";
+import { cargarImagenPerfil, deleteImgProfile, editImgProfile } from "../redux/Trunks/userTrunk";
+import Cookies from "js-cookie";
+import { encryptSessionIndicator } from "../security/Encr_decp";
+import { LimpiarURL } from "../redux/Slides/userSlice";
 
 interface UserProfileProps {
     prof_pic?: string;
@@ -15,16 +18,18 @@ interface UserProfileProps {
 
 const ChangeProfileIMG: React.FC<UserProfileProps> = ({ prof_pic, token, id }) => {
     const [image, setImage] = useState<File | null>(null); // Imagen de perfil seleccionada.
-    const [imagePath, /*setImagePath*/] = useState<string | undefined>(prof_pic); // Ruta de la imagen antigua.
+    const [imagePath, setImagePath] = useState<string | undefined>(prof_pic); // Ruta de la imagen antigua.
     const [preview, setPreview] = useState<string | null>(null); // Vista previa de la imagen.
     const [isModalOpen, setIsModalOpen] = useState(false); // Controla si el modal está abierto.
     const [modalAction, setModalAction] = useState<"delete" | "change" | null>(null); // Define la acción a confirmar.
     const [mensajeModal, setMensajeModal] = useState<string | undefined>(undefined); // Mensaje del modal.
     const [isOpenLoading, setIsOpenLoading] = useState(false); // Modal de carga.
 
-    let imgPathNew = '';
-    console.log(imagePath);
+    const { currentUser } = useSelector(
+        (state: RootState) => state.users
+    );
 
+    let imgPathNew = '';
 
     const dispatch = useDispatch<AppDispatch>(); // Medio que se utiliza en el desapachador
 
@@ -42,30 +47,54 @@ const ChangeProfileIMG: React.FC<UserProfileProps> = ({ prof_pic, token, id }) =
     // Maneja la confirmación de las acciones del modal.
     const handleConfirm = async () => {
         if (modalAction === "delete") {
-            setImage(null);
-            setPreview(null);
-            // setImagePath(undefined); // Borra la imagen
-            if (fileInputRef.current) {
-                fileInputRef.current.value = ""; // Resetea el valor del input de tipo file
+            if (preview) {
+                // Borrar solo la vista previa y resetear el input
+                if (fileInputRef.current) {
+                    fileInputRef.current.value = ""; // Limpia el input
+                }
+                setImage(null);
+                setPreview(null);
+                closeModal();
+                return; // Salimos antes de ejecutar el borrado en el servidor
             }
+            
             if (imagePath) {
                 setIsOpenLoading(true);
                 setMensajeModal('Borrando la imagen...');
-                const resultAction = await dispatch(deleteImgProfile({id_usuario: id, token: token}));
+                const resultAction = await dispatch(deleteImgProfile({ id_usuario: id, token: token }));
                 // Manejor de la respuesta de la subida
                 if (deleteImgProfile.fulfilled.match(resultAction)) {
                     const uploadCorrect = resultAction.payload;
-                    console.log(uploadCorrect);
 
                     if (!uploadCorrect) {
                         throw new Error('Tuvimos problemas para subir la imagen, inténtalo más tarde');
                     }
+                    dispatch(LimpiarURL()); // Limpia la URL
+                    setImagePath(undefined); // Borra la imagen
+                    const updatedUserCookie = {
+                        id_usuario: currentUser?.id_usuario,
+                        nick: currentUser?.nick,
+                        token: currentUser?.token,
+                        url_prof_pic: imgPathNew,
+                    };
+
+                    // Encriptar los datos
+                    const encryptedSession = await encryptSessionIndicator(JSON.stringify(updatedUserCookie));
+
+                    // Guardar la cookie encriptada
+                    Cookies.set("isLoggedIn", encryptedSession, {
+                        expires: new Date(new Date().getTime() + 60 * 60 * 1000), // Expira en 1 hora
+                        secure: true, // Solo en conexiones HTTPS
+                        sameSite: "Strict", // Prevenir ataques CSRF
+                    });
 
                 } else if (deleteImgProfile.rejected.match(resultAction)) {
                     const errorMessage = resultAction?.payload as string || 'Error al subir la imagen.';
                     throw new Error(errorMessage);
                 }
                 setIsOpenLoading(false);
+                // Dispatch del thunk para cargar la imagen de perfil
+                dispatch(cargarImagenPerfil({ url_prof_pic: imgPathNew, token: token }));
             }
 
         } else if (modalAction === "change") {
@@ -83,7 +112,6 @@ const ChangeProfileIMG: React.FC<UserProfileProps> = ({ prof_pic, token, id }) =
                         throw new Error('Tuvimos problemas para subir la imagen, inténtalo más tarde');
                     }
 
-
                     imgPathNew = uploadCorrect;
                 } else if (subirImagenPerfil.rejected.match(resultAction)) {
                     const errorMessage = resultAction?.payload as string || 'Error al subir la imagen.';
@@ -95,18 +123,42 @@ const ChangeProfileIMG: React.FC<UserProfileProps> = ({ prof_pic, token, id }) =
             // Manejor de la respuesta del cambio
             if (editImgProfile.fulfilled.match(resultAction)) {
                 const uploadCorrect = resultAction.payload;
-
+                
                 if (!uploadCorrect) {
                     throw new Error('Tuvimos problemas para cambiar la imagen, inténtalo más tarde');
                 }
+                
+                dispatch(LimpiarURL()); // Limpia la URL
+                const updatedUserCookie = {
+                    id_usuario: currentUser?.id_usuario,
+                    nick: currentUser?.nick,
+                    token: currentUser?.token,
+                    url_prof_pic: imgPathNew,
+                };
+
+                // Encriptar los datos
+                const encryptedSession = await encryptSessionIndicator(JSON.stringify(updatedUserCookie));
+
+                // Guardar la cookie encriptada
+                Cookies.set("isLoggedIn", encryptedSession, {
+                    expires: new Date(new Date().getTime() + 60 * 60 * 1000), // Expira en 1 hora
+                    secure: true, // Solo en conexiones HTTPS
+                    sameSite: "Strict", // Prevenir ataques CSRF
+                });
 
             } else if (subirImagenPerfil.rejected.match(resultAction)) {
                 const errorMessage = resultAction?.payload as string || 'Error al cambiar la imagen.';
                 throw new Error(errorMessage);
             }
+            dispatch(cargarImagenPerfil({ url_prof_pic: imgPathNew, token: token }));
             setIsOpenLoading(false);
         }
         closeModal();
+        setImage(null);
+        setPreview(null);
+        if (fileInputRef.current) {
+            fileInputRef.current.value = ""; // Resetea el valor del input de tipo file
+        }
     };
 
     const openModal = (action: "delete" | "change") => {
@@ -121,40 +173,48 @@ const ChangeProfileIMG: React.FC<UserProfileProps> = ({ prof_pic, token, id }) =
 
     return (
         <>
-            <form>
-                <p>Cambiar/ Borrar / Subir imagen de perfil</p>
+            <form className="profile-form">
+                <p className="profile-form-title">Editar imagen de perfil</p>
                 <input
                     type="file"
+                    className="profile-form-input"
                     accept="image/jpeg, image/png, image/webp, image/gif"
                     onChange={handleImageChange}
                     ref={fileInputRef} // Asigna la referencia al input
                 />
 
-                <div style={{ display: 'flex', flexDirection: 'column', alignItems: 'center', height: '200px' }}>
-                    <h3>Vista Previa:</h3>
+                <div className="profile-preview-container">
                     {preview && (
-                        <div>
-                            <img src={preview} alt="Vista previa" style={{ maxWidth: '125px', maxHeight: '125px', borderRadius: '50%' }} />
-                            <div className="btn-container">
+                        <div className="profile-preview">
+                            <h3 className="profile-preview-title">Vista Previa:</h3>
+                            <img
+                                src={preview}
+                                alt="Vista previa"
+                                className="profile-preview-image"
+                            />
+                            <div className="profile-buttons-container">
                                 <button
                                     type="button"
-                                    className="button-signup"
+                                    className="profile-button profile-button-change"
                                     onClick={() => openModal("change")}
                                 >
-                                    Cambiar la imagen
+                                    Confirmar imagen
                                 </button>
                             </div>
                         </div>
                     )}
                 </div>
-                <div>
+                
+                <div className="profile-buttons-container">
                     <button
                         type="button"
-                        className="button-signup"
+                        className="profile-button profile-button-delete"
                         onClick={() => openModal("delete")}
                         disabled={!imagePath && !preview} // Desactiva si no hay ruta ni vista previa
                     >
-                        {imagePath ? 'Borrar imagen antigua' : 'Borrar imagen seleccionada'}
+                        {imagePath && !preview
+                            ? "Borrar imagen actual"
+                            : "Descartar imagen seleccionada"}
                     </button>
                 </div>
             </form>
